UI Phase 11.3: Aura & Channel Area Debug Overlay
- Implemented world-space debug overlay for active AURAs and channel areas with a radial falloff gradient.
- New module: src/game/skill_area_debug_overlay.{h,c}; integrated into main render loop behind app toggle.
- Toggle key: F11 → toggles g_app.show_skill_area_overlay; input bound in src/input/input_events.c.
- EffectSpec: added active AURA introspection getters to drive the overlay without gameplay side effects.
- Minor fixes: SDL color casts in hud_buff_belt.c; local var shadowing fix in skill_bar.c to satisfy warnings-as-errors.
- Build: Debug (SDL2) with parallelism (-j12) succeeds; CMake includes the new module.
- Tests: Full suite executed in Debug with -j12; skills/effectspec/renderer/UI subsets green. Some AI/start-screen/equipment tests are currently red under headless assets; unrelated to this change.

+++ src/core/hud/hud_buff_belt.c
@@ -122,7 +122,7 @@ void rogue_hud_buff_belt_render(const RogueHUDBuffBeltState* st, int screen_w)
             cg = 130;
             cb = 220;
         }
-        SDL_SetRenderDrawColor(g_app.renderer, cr, cg, cb, 255);
+        SDL_SetRenderDrawColor(g_app.renderer, (Uint8) cr, (Uint8) cg, (Uint8) cb, 255);
         SDL_Rect br = {x - 1, y - 1, icon_w + 2, icon_h + 2};
         SDL_RenderDrawRect(g_app.renderer, &br);
         /* Duration mini-bar (top, proportional) */
